requirements.txt
torch==2.2.1
diffusers==0.27.2
transformers==4.40.2
xformers==0.0.25
Pillow==10.0.1
accelerate==0.28.0
huggingface_hub==0.23.0
safetensors
peft  # For LoRA
numpy<2
